Pulled IEFileType (wp-specific) out of xap base classes, made ints, do explicit casting in derived class members which implement these virtual functions.  I tried to have one level of casting (in the first level of entry into these derived classes) so that IEFileType can be properly used inside the wp tree everywhere but those class members.
This should be the last of the IEFileTypes remaining in XAP code, so
that new applications can use the FileOpenSaveAs dialog, the XAP frames,
and XAP documents with their own document types.


git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@1390 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/wp/ap/xp/ap_EditMethods.cpp
@@ -1124,7 +1124,7 @@ Defun1(fileSaveAs)
 
 	UT_DEBUGMSG(("fileSaveAs: saving as [%s]\n",pNewFile));
 
-	UT_Bool bSaved = pAV_View->cmdSaveAs(pNewFile,ieft);
+	UT_Bool bSaved = pAV_View->cmdSaveAs(pNewFile,(int) ieft);
 
 	if (!bSaved)
 	{
